# Blazor_Wasm/OpenAPIs/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Core_API",
    "version": "1.0"
  },
  "paths": {
    "/getall": {
      "get": {
        "tags": [
          "Department"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Department"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/Department/getone/{id}": {
      "get": {
        "tags": [
          "Department"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Department"
                }
              }
            }
          }
        }
      }
    },
    "/createone": {
      "post": {
        "tags": [
          "Department"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Department"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Department"
                }
              }
            }
          }
        }
      }
    },
    "/update/{id}": {
      "put": {
        "tags": [
          "Department"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Department"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Department"
                }
              }
            }
          }
        }
      }
    },
    "/delete/{id}": {
      "delete": {
        "tags": [
          "Department"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Department": {
        "type": "object",
        "properties": {
          "DeptNo": {
            "type": "integer",
            "format": "int32"
          },
          "DeptName": {
            "type": "string",
            "nullable": true
          },
          "Capacity": {
            "type": "integer",
            "format": "int32"
          },
          "Location": {
            "type": "string",
            "nullable": true
          },
          "Employees": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Employee"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Employee": {
        "type": "object",
        "properties": {
          "EmpNo": {
            "type": "integer",
            "format": "int32"
          },
          "EmpName": {
            "type": "string",
            "nullable": true
          },
          "Designation": {
            "type": "string",
            "nullable": true
          },
          "Salary": {
            "type": "integer",
            "format": "int32"
          },
          "DeptNo": {
            "type": "integer",
            "format": "int32"
          },
          "DeptNoNavigation": {
            "$ref": "#/components/schemas/Department"
          }
        },
        "additionalProperties": false
      }
    }
  }
}